Change from Linus

+++ arch/powerpc/include/asm/uaccess.h
@@ -434,8 +434,8 @@ copy_mc_to_user(void __user *to, const void *from, unsigned long n)
 }
 #endif
 
-extern long __copy_from_user_flushcache(void *dst, const void __user *src,
-		unsigned size);
+extern unsigned long __copy_from_user_flushcache(void *dst, const void __user *src,
+		unsigned long size);
 
 static __must_check __always_inline bool __user_access_begin(const void __user *ptr, size_t len,
 							     unsigned long dir)
