+++ arch/x86/include/asm/uaccess_64.h
@@ -147,12 +147,11 @@ raw_copy_to_user(void __user *dst, const void *src, unsigned long size)
 	return copy_user_generic((__force void *)dst, src, size);
 }
 
-extern long __copy_user_nocache(void *dst, const void __user *src, unsigned size);
-extern long __copy_user_flushcache(void *dst, const void __user *src, unsigned size);
+extern unsigned long __copy_user_nocache(void *dst, const void __user *src, unsigned long size);
+extern unsigned long __copy_user_flushcache(void *dst, const void __user *src, unsigned long size);
 
-static inline int
-__copy_from_user_inatomic_nocache(void *dst, const void __user *src,
-				  unsigned size)
+static inline unsigned long
+__copy_from_user_inatomic_nocache(void *dst, const void __user *src, unsigned long size)
 {
 	long ret;
 	kasan_check_write(dst, size);
@@ -162,8 +161,8 @@ __copy_from_user_inatomic_nocache(void *dst, const void __user *src,
 	return ret;
 }
 
-static inline int
-__copy_from_user_flushcache(void *dst, const void __user *src, unsigned size)
+static inline unsigned long
+__copy_from_user_flushcache(void *dst, const void __user *src, unsigned long size)
 {
 	kasan_check_write(dst, size);
 	return __copy_user_flushcache(dst, src, size);
